Update modeling.py
for the BertLMHeadModel class, we can process the batch with or without Ngram module according to the is_Ngarm argument. When is_Ngram is True, we add the bertNgramEmbedding module for the input_Ngram_ids. And we add the attention module, then we add the Ngram_embedding for the Ngrams and the embedding for the tokens.